fix: remove adaptive card codelens provider
