+++ packages/vscode-extension/src/handlers.ts
@@ -12,7 +12,6 @@ import * as fs from "fs-extra";
 import * as path from "path";
 import * as util from "util";
 import * as uuid from "uuid";
-import { glob } from "glob";
 import * as vscode from "vscode";
 
 import {
@@ -1967,56 +1966,9 @@ export async function installAdaptiveCardExt(
   return Promise.resolve(ok(null));
 }
 
-function isAdaptiveCard(content: string): boolean {
-  const pattern = /"type"\s*:\s*"AdaptiveCard"/;
-  return pattern.test(content);
-}
-
 export function acpInstalled(): boolean {
   const extension = vscode.extensions.getExtension(acExtId);
   return !!extension;
-}
-
-export async function hasAdaptiveCardInWorkspace(): Promise<boolean> {
-  // Skip large files which are unlikely to be adaptive cards to prevent performance impact.
-  const fileSizeLimit = 1024 * 1024;
-
-  if (globalVariables.workspaceUri) {
-    const files = await glob(globalVariables.workspaceUri.path + "/**/*.json", {
-      ignore: ["**/node_modules/**", "./node_modules/**"],
-    });
-    for (const file of files) {
-      let content = "";
-      let fd = -1;
-      try {
-        fd = await fs.open(file, "r");
-        const stat = await fs.fstat(fd);
-        // limit file size to prevent performance impact
-        if (stat.size > fileSizeLimit) {
-          continue;
-        }
-
-        // avoid security issue
-        // https://github.com/OfficeDev/TeamsFx/security/code-scanning/2664
-        const buffer = new Uint8Array(fileSizeLimit);
-        const { bytesRead } = await fs.read(fd, buffer, 0, buffer.byteLength, 0);
-        content = new TextDecoder().decode(buffer.slice(0, bytesRead));
-      } catch (e) {
-        // skip invalid files
-        continue;
-      } finally {
-        if (fd >= 0) {
-          fs.close(fd);
-        }
-      }
-
-      if (isAdaptiveCard(content)) {
-        return true;
-      }
-    }
-  }
-
-  return false;
 }
 
 export async function openPreviewAadFile(args: any[]): Promise<Result<any, FxError>> {
