+++ packages/vscode-extension/src/utils/commonUtils.ts
@@ -21,6 +21,7 @@ import { getV3TeamsAppId } from "../debug/commonUtils";
 import * as globalVariables from "../globalVariables";
 import { core } from "../handlers";
 import { TelemetryProperty, TelemetryTriggerFrom } from "../telemetry/extTelemetryEvents";
+import { glob } from "glob";
 
 export function getPackageVersion(versionStr: string): string {
   if (versionStr.includes("alpha")) {
@@ -397,4 +398,51 @@ export function getTriggerFromProperty(args?: any[]) {
     default:
       return { [TelemetryProperty.TriggerFrom]: TelemetryTriggerFrom.Unknow };
   }
+}
+
+export async function hasAdaptiveCardInWorkspace(): Promise<boolean> {
+  // Skip large files which are unlikely to be adaptive cards to prevent performance impact.
+  const fileSizeLimit = 1024 * 1024;
+
+  if (globalVariables.workspaceUri) {
+    const files = await glob(globalVariables.workspaceUri.path + "/**/*.json", {
+      ignore: ["**/node_modules/**", "./node_modules/**"],
+    });
+    for (const file of files) {
+      let content = "";
+      let fd = -1;
+      try {
+        fd = await fs.open(file, "r");
+        const stat = await fs.fstat(fd);
+        // limit file size to prevent performance impact
+        if (stat.size > fileSizeLimit) {
+          continue;
+        }
+
+        // avoid security issue
+        // https://github.com/OfficeDev/TeamsFx/security/code-scanning/2664
+        const buffer = new Uint8Array(fileSizeLimit);
+        const { bytesRead } = await fs.read(fd, buffer, 0, buffer.byteLength, 0);
+        content = new TextDecoder().decode(buffer.slice(0, bytesRead));
+      } catch (e) {
+        // skip invalid files
+        continue;
+      } finally {
+        if (fd >= 0) {
+          fs.close(fd).catch(() => {});
+        }
+      }
+
+      if (isAdaptiveCard(content)) {
+        return true;
+      }
+    }
+  }
+
+  return false;
+}
+
+function isAdaptiveCard(content: string): boolean {
+  const pattern = /"type"\s*:\s*"AdaptiveCard"/;
+  return pattern.test(content);
 }
