+++ packages/vscode-extension/test/extension/handlers.test.ts
@@ -2556,75 +2556,6 @@ describe("autoOpenProjectHandler", () => {
     chai.assert.isTrue(executeCommandStub.calledOnce);
   });
 
-  describe("hasAdaptiveCardInWorkspace()", () => {
-    afterEach(() => {
-      mockfs.restore();
-    });
-
-    it("no workspace", async () => {
-      sandbox.stub(globalVariables, "workspaceUri").value(undefined);
-
-      const result = await handlers.hasAdaptiveCardInWorkspace();
-
-      chai.assert.isFalse(result);
-    });
-
-    it("happy path", async () => {
-      sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("/test"));
-      mockfs({
-        "/test/card.json": JSON.stringify({
-          $schema: "http://adaptivecards.io/schemas/adaptive-card.json",
-          type: "AdaptiveCard",
-          version: "1.5",
-          actions: [
-            {
-              type: "Action.OpenUrl",
-              title: "More Info",
-              url: "https://example.com",
-            },
-          ],
-        }),
-      });
-
-      const result = await handlers.hasAdaptiveCardInWorkspace();
-
-      chai.assert.isTrue(result);
-    });
-
-    it("hasAdaptiveCardInWorkspace() no adaptive card file", async () => {
-      sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("/test"));
-      mockfs({
-        "/test/card.json": JSON.stringify({ hello: "world" }),
-      });
-
-      const result = await handlers.hasAdaptiveCardInWorkspace();
-
-      chai.assert.isFalse(result);
-    });
-
-    it("hasAdaptiveCardInWorkspace() very large adaptive card file", async () => {
-      sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("/test"));
-      mockfs({
-        "/test/card.json": JSON.stringify({
-          $schema: "http://adaptivecards.io/schemas/adaptive-card.json",
-          type: "AdaptiveCard",
-          version: "1.5",
-          actions: [
-            {
-              type: "Action.OpenUrl",
-              title: "a".repeat(1024 * 1024 + 10),
-              url: "https://example.com",
-            },
-          ],
-        }),
-      });
-
-      const result = await handlers.hasAdaptiveCardInWorkspace();
-
-      chai.assert.isFalse(result);
-    });
-  });
-
   describe("acpInstalled()", () => {
     afterEach(() => {
       mockfs.restore();
