+++ packages/vscode-extension/test/extension/treeview/treeViewManager.test.ts
@@ -2,12 +2,9 @@ import * as chai from "chai";
 import * as sinon from "sinon";
 import * as vscode from "vscode";
 
-import { AdaptiveCardCodeLensProvider } from "../../../src/codeLensProvider";
 import * as globalVariables from "../../../src/globalVariables";
 import { CommandsTreeViewProvider } from "../../../src/treeview/commandsTreeViewProvider";
 import treeViewManager from "../../../src/treeview/treeViewManager";
-import { manifestUtils } from "@microsoft/teamsfx-core";
-import { TeamsAppManifest, ok } from "@microsoft/teamsfx-api";
 
 describe("TreeViewManager", () => {
   const sandbox = sinon.createSandbox();
@@ -54,73 +51,6 @@ describe("TreeViewManager", () => {
     chai.assert.equal(setStatusStub.callCount, 2);
   });
 
-  it("updateTreeViewsByContent has adaptive cards", async () => {
-    sandbox
-      .stub(AdaptiveCardCodeLensProvider, "detectedAdaptiveCards")
-      .returns(Promise.resolve(true));
-    sandbox.stub(manifestUtils, "readAppManifest").resolves(ok({} as TeamsAppManifest));
-    sandbox.stub(manifestUtils, "getCapabilities").returns(["tab"]);
-
-    treeViewManager.registerTreeViews({
-      subscriptions: [],
-    } as unknown as vscode.ExtensionContext);
-    const utilityTreeviewProvider = treeViewManager.getTreeView(
-      "teamsfx-utility"
-    ) as CommandsTreeViewProvider;
-
-    const commands = utilityTreeviewProvider.getCommands();
-    chai.assert.equal(commands.length, 3);
-
-    await treeViewManager.updateTreeViewsByContent();
-
-    chai.assert.equal(commands.length, 4);
-  });
-
-  it("updateTreeViewsByContent that removes project related commands", async () => {
-    sandbox
-      .stub(AdaptiveCardCodeLensProvider, "detectedAdaptiveCards")
-      .returns(Promise.resolve(true));
-    sandbox.stub(globalVariables, "isSPFxProject").value(false);
-    sandbox.stub(manifestUtils, "readAppManifest").resolves(ok({} as TeamsAppManifest));
-    sandbox.stub(manifestUtils, "getCapabilities").returns(["tab"]);
-
-    treeViewManager.registerTreeViews({
-      subscriptions: [],
-    } as unknown as vscode.ExtensionContext);
-    const developmentTreeviewProvider = treeViewManager.getTreeView(
-      "teamsfx-development"
-    ) as CommandsTreeViewProvider;
-
-    const commands = developmentTreeviewProvider.getCommands();
-    chai.assert.equal(commands.length, 4);
-
-    await treeViewManager.updateTreeViewsByContent(true);
-
-    chai.assert.equal(commands.length, 3);
-  });
-
-  it("updateTreeViewsByContent that is not teams app", async () => {
-    sandbox
-      .stub(AdaptiveCardCodeLensProvider, "detectedAdaptiveCards")
-      .returns(Promise.resolve(false));
-    sandbox.stub(manifestUtils, "readAppManifest").resolves(ok({} as TeamsAppManifest));
-    sandbox.stub(manifestUtils, "getCapabilities").returns([]);
-
-    treeViewManager.registerTreeViews({
-      subscriptions: [],
-    } as unknown as vscode.ExtensionContext);
-    const utilityTreeviewProvider = treeViewManager.getTreeView(
-      "teamsfx-utility"
-    ) as CommandsTreeViewProvider;
-
-    const commands = utilityTreeviewProvider.getCommands();
-    chai.assert.equal(commands.length, 3);
-
-    await treeViewManager.updateTreeViewsByContent();
-
-    chai.assert.equal(commands.length, 2);
-  });
-
   it("updateTreeViewsOnSPFxChanged", () => {
     sandbox.stub(globalVariables, "isSPFxProject").value(false);
     treeViewManager.registerTreeViews({
@@ -137,25 +67,5 @@ describe("TreeViewManager", () => {
     treeViewManager.updateTreeViewsOnSPFxChanged();
 
     chai.assert.equal(commands.length, 5);
-  });
-
-  it("should detected adaptive card", async () => {
-    const provider = new AdaptiveCardCodeLensProvider();
-    const res = await provider.provideCodeLenses({
-      getText(range?: Range): string {
-        return '"$schema": "http://adaptivecards.io/schemas/adaptive-card.json",';
-      },
-    } as vscode.TextDocument);
-    chai.assert.isNotEmpty(res);
-  });
-
-  it("adaptive card is not detected", async () => {
-    const provider = new AdaptiveCardCodeLensProvider();
-    const res = await provider.provideCodeLenses({
-      getText(range?: Range): string {
-        return "";
-      },
-    } as vscode.TextDocument);
-    chai.assert.equal(res?.length, 0);
   });
 });
